requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
exceptiongroup==1.3.0
fastapi==0.119.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pip==22.0.2
pydantic==2.12.2
pydantic_core==2.41.4
python-decouple==3.8
requests==2.32.5
setuptools==59.6.0
sniffio==1.3.1
starlette==0.48.0
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.37.0
Werkzeug==3.1.3
